expose getAccount
