+++ src/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from "vitest";
-import { startSiwf } from "./index.js";
+import { getGatewayAccount, startSiwf } from "./index.js";
 import { mockGatewayFetchFactory } from "../test-mocks/mockGatewayFetchFn";
 import {
   mockChainInfoResponse,
@@ -14,7 +14,8 @@ import {
 } from "../test-mocks/consts";
 import { decodeSignedRequest } from "@projectlibertylabs/siwf";
 import { AccountResponse } from "./gateway-types";
-import { EIP712, MsaCreationCallbackFn } from "./types";
+import { EIP712, GatewayFetchFn, MsaCreationCallbackFn } from "./types";
+import { GatewayFetchError } from "./error-types";
 
 const providerControlKey = decodeSignedRequest(mockProviderEncodedRequest)
   .requestedSignatures.publicKey.encodedValue;
@@ -183,5 +184,57 @@ describe("Basic startSiwf test", () => {
         mockReturningUserAccountResponse,
       );
     });
+  });
+});
+
+describe("getAccountForAccountId", () => {
+  it("returns info for an existing user", async () => {
+    const body: AccountResponse = { msaId: "2" };
+    const fetchFn: GatewayFetchFn = async (_method, _path) => {
+      return new Response(JSON.stringify(body), { status: 200 });
+    };
+    const address = "0xf24FF3a9CF04c71Dbc94D0b566f7A27B94566cac";
+
+    const result = await getGatewayAccount(fetchFn, address);
+
+    expect(result).toStrictEqual(body);
+  });
+  it("returns `null` when user not found", async () => {
+    const fetchFn: GatewayFetchFn = async (_method, _path) => {
+      return new Response(null, { status: 404 });
+    };
+    const address = "0xf24FF3a9CF04c71Dbc94D0b566f7A27B94566cac";
+
+    const result = await getGatewayAccount(fetchFn, address);
+
+    expect(result).toStrictEqual(null);
+  });
+  it("throws when request is malformed", async () => {
+    const fetchFn: GatewayFetchFn = async (_method, _path) => {
+      return new Response(null, { status: 405 });
+    };
+    const address = "#@41i9=/?&8";
+
+    try {
+      await getGatewayAccount(fetchFn, address);
+      expect.fail("No error thrown");
+    } catch (err) {
+      expect(err).toBeInstanceOf(GatewayFetchError);
+      expect((err as GatewayFetchError).response.status).toBe(405);
+    }
+  });
+  it("throws when server encounters an error", async () => {
+    const fetchFn: GatewayFetchFn = async (_method, _path) => {
+      return new Response("{stacktrace: '...'}", { status: 500 });
+    };
+    const address = "0xf24FF3a9CF04c71Dbc94D0b566f7A27B94566cac";
+
+    try {
+      await getGatewayAccount(fetchFn, address);
+      expect.fail("No error thrown");
+    } catch (err) {
+      expect(err).toBeInstanceOf(GatewayFetchError);
+      expect((err as GatewayFetchError).response.status).toBe(500);
+    }
   });
 });
