+++ src/index.ts
@@ -16,7 +16,6 @@ import {
   createSignedGraphKeyPayload,
 } from "./helpers/payloads.js";
 import {
-  getGatewayAccount,
   getGatewayChainInfo,
   pollForAccount,
   postGatewaySiwf,
@@ -33,6 +32,7 @@ import {
   CreateLoginSiwfResponseArguments,
   createSignInSiwfResponse,
 } from "./helpers/siwf";
+import { GatewayFetchError } from "./error-types";
 
 const PAYLOAD_EXPIRATION_DELTA = 90; // Matches frequency access config
 
@@ -229,5 +229,36 @@ export async function startSiwf(
       decodedSiwfSignedRequest,
       chainInfo,
     );
+  }
+}
+
+/**
+ * Fetches a user's account information (if present) from Gateway Services
+ *
+ * @param gatewayFetchFn Callback for performing request to gateway services
+ * @param userAddress - the public key of the user who wishes to sign in
+ * @returns An 'account response' when the user's account exists, and `null` otherwise
+ * @throws `GatewayFetchError` when the request fails
+ */
+export async function getGatewayAccount(
+  gatewayFetchFn: GatewayFetchFn,
+  userAddress: string,
+): Promise<AccountResponse | null> {
+  const response = await gatewayFetchFn(
+    "GET",
+    `/v1/accounts/account/${userAddress}`,
+  );
+  if (response.ok) {
+    return (await response.json()) as AccountResponse;
+  } else {
+    switch (response.status) {
+      case 404:
+        return null; // The user does not (yet) exist on chain
+      default:
+        throw new GatewayFetchError(
+          "Failed GatewayFetchFn for GET Account",
+          response,
+        );
+    }
   }
 }
